Go: Expose writer indexes as public (#86)
Expose the chunk, attachment, message indexes as well as statistics as
public fields on the writer. These are useful if you want to maintain an
external index of the files you write. It saves the need to reread them
for the index data after writing, which would require a seek or range
request.