Swap arg order for bag conversion functions
Swaps the argument order for the bag to mcap conversion function so that
the writer is the first argument, which is more conventional in Go. Also
addresses two issues with the CLI tool - the cat command now indicates
where the input file goes in its help output, and the start and end
flags are now start-secs and end-secs.