Change lexer's interface to byte arrays
Removes the "Token" type from the go library, in favor of passing around
tuples of (type, bytes). This is enabled by changing the Next() methods
of the readers and lexer to accept an optional slice of bytes from which
the outputs may be sliced, to reduce allocations. Consequently the
consumer of the lexer no longer needs to worry about consuming the
indicated number of bytes from the exposed reader. The lexer will handle
that for them.