Go: Address ros2 db3 conversion bug for multiple input paths
Prior to this commit, if the ros2 conversion function failed to find a
package it was looking for, it would break out of a loop immediately and
error on a check that all requested packages were found. This did not
properly account for the possibility of the package existing in a later
path element. The new code continues instead of breaking so that the
schema may be found in another location if available.