Revert "Go: do not write chunk indexes for chunks without messages (#257)" (#260)
This reverts commit a7b663816a3528e55c8c238ce2a46f2f03245524.

The spec actually allows for chunk indexes for chunks without messages,
although the concept is not really sensical. I think a change will need
to be made, probably to include the message count in the chunk index
records, but I don't think this approach is was the right one anymore.