Add CLI support for remote files
Adds remote file support for the "info", "cat", and "doctor" commands to
exhibit functionality of the index on remote files. Leaves out "convert"
for now to skip refactoring some of the magic number checking.

Currently works for GCS only: filenames that start with gs:// will be
interpreted as GCS files and accessed using ADC credentials. To extend
to other object stores we can implement seekable readers on their
objects as well.