Remove replace directives from ros library and cli
Prior to this commit the subprojects referenced each other to allow
simultaneous changes, however this breaks the installation of the CLI
tool via go get github.com/...

This commit removes the mechanism that allowed this. Going forward
separate PRs will be required for updates to different libraries. Leaves
the directives in place for the conformance tests, since they are
necessary there.