Go: Add support for sorting serialized chunks
Adds a SortChunkMessages option to the go writer, defaulted to false,
that sorts the messages and message index records in chunks prior to
writing to disk.

The intent of the functionality is to allow readers of files produced in
this way to execute streaming reads on compressed chunks and stop once
the first out-of-range message is seen.

The sort used for the index records is insertion sort, under the
assumption messages will be mostly/almost-always time-ordered to start.
The sort of the chunk mirrors the swaps executed by the insertion sort,
staging copies through the w.msg buffer. Since the same buffer was used
to stage writes to the chunk, it is guaranteed to have sufficient room
without additional allocations.