Perform single seek to implicated chunk range
Refactors the indexed playback strategy so that the message indexes for
a chunk are read in the same read as the chunk itself, not prior. In
combination with the prior commit, to skip noop seeks in the remote
reader, this allows us to play back an mcap file that has only chunks
and message index records at the top level, with three seeks.