Go: reduce seeking in indexed playback (#344)
* Read message index section in single pass

Prior to this commit, the indexed reader did a seek for each message
index record required by the query. On remote connections it'll be much
faster to read the whole section and discard irrelevant records. On
local systems it shouldn't matter.

* Don't seek remote reader if already in position

If the current offset of the reader matches the one being seeked to,
keep using the same reader instead of opening a new request.

* Perform single seek to implicated chunk range

Refactors the indexed playback strategy so that the message indexes for
a chunk are read in the same read as the chunk itself, not prior. In
combination with the prior commit, to skip noop seeks in the remote
reader, this allows us to play back an mcap file that has only chunks
and message index records at the top level, with three seeks.