CLI: Fix zero-valued frequencies (#381)
Fixes a bug introduced recently that caused frequencies in `info` output
to be zero due to unintentional shadowing of a global variable used for
the cat command.

Use of global variables is an artifact of cobra's CLI command code
generation. We can get away from that with a bit of refactoring but
haven't done so yet. As a stopgap, this commit also introduces a
convention of prefixing the globals with the name of the related
command.