python: fix zstd decompression (#390)
**Public-Facing Changes**
Fixed zstd decompression in the Python reader.


**Description**
Fixes #389
The `zstd` package did not allow us to pass the uncompressed size as a parameter to the decompress function. The `zstandard` package does. This is required when the writer used a streaming compression API.

Tests are not added because the Python writer uses the one-shot zstd compress API, so does not trigger this issue in the reader (which only occurs when the compressed frame does not include a size value, i.e. when a streaming compression API was used).