Add `reverse` option to Mcap0IndexedReader#readMessages (#391)
Add a new `reverse` option for the `readMessages` method of Mcap0IndexedReader. The reverse option emits messages in reverse logTime order (i.e. older messages first). This is useful when to find the latest available message on all requested topics at some arbitrary time within a file.

`startTime` and `endTime` arguments are _not_ reversed when using the `reverse` option. Only the order of the messages produced from the iterator.