go/mcap: fix bug writing attachment indices
There was a bug where WriteAttachmentIndex would fail to
pre-allocate 8 bytes for one of its uint64 timestamp fields.

This would cause the write buffer to be too short to fit the
final field (the content type string). In our existing test cases,
the content type was more than 8 chars long, and the effect was
to truncate the content type string by 8 chars. However,
in implementing `mcap filter`, i created a test case with a 0-length
content type string, which caused a segfault.