python: support packages can use indexed reader (#519)
Replaces ROS1 and Protobuf support API for decoding messages.

The previous support libraries used the StreamReader to decode messages, which does not take advantage of the index to search for messages by topic or timestamp. This PR replaces those decoding classes with an iterator adapter, which consumes the `McapReader.iter_messages()` iterator. This allows them to take advantage of the index as well.

Fixes https://github.com/foxglove/private-issues/issues/47