python: add option to order read messages by log time (#528)
**Public-Facing Changes**
Adds an option to `iter_messages` to read messages in log time order.


<!-- describe any changes to the public interface or APIs, or write "None" -->

**Description**
On indexed, seekable MCAP sources, this implementation uses a priority queue to order messages in ascending `log_time` order. Chunks are decompressed when their `message_start_time` is the earliest in the queue.

On unindexed or non-seekable MCAP sources, this implementation loads all message records into memory and sort them by log time. This uses a lot of memory, but for non-seekable sources this is not avoidable. For non-indexed but seekable MCAP sources, this could be improved - we could read the log times into memory up-front but only read the content on-demand.
<!-- describe what has changed, and motivation behind those changes -->

Fixes: #529