go: read in log time order with overlapping chunks (#534)
**Public-Facing Changes**
Adds a `reverse` argument to McapReader.Messages(), letting a user read in reverse order.

Indexed MCAP reads will now return messages in log time order, even if there are chunks with overlapping time ranges.


# Description
This PR changes the approach to reading messages using the indexed reader. 

## Before
The previous implementation had one active chunk, and all messages were read out of it in ascending log time order before moving on to the next chunk. If two chunks are present in an MCAP with overlapping time ranges, this would cause messages to be read out-of-order.

## After
This PR changes to a scheme where there is one heap containing chunk indexes and message index entries that match the search criteria. 

When a chunk index reaches the top of the heap, the corresponding chunk is decompressed and all of its message index entries are pushed to the heap.

Message index entries in the heap are stored alongside a pointer to the underlying decompressed chunk. There is no "owning" pointer to the decompressed chunk data, so we now rely on the garbage collector to clean up the decompressed chunk once all message index entries are consumed.

Additionally, since it's easy enough, this PR adds a flag to iterate through the MCAP in reverse time order. This can be used to search for the first message on a topic before a given timestamp.
Fixes #533.