CLI: Support eloquent bagfiles (#570)
* CLI: Open writer after scanning topics in ros2 conversion

Prior to this commit, we would open the writer before reading topics out
of the input db3 file. If the topics read failed, this resulted in a
garbage output file.

* Support eloquent bagfiles

In eloquent, there is no offered_qos_profiles column in the topics
table. Accomodate this, and only insert it into channel metadata if it
is present as a column.

* Add cspell.json

* Add test file