Add support for adding metadata/attachments (#614)
Adds support for adding and getting metadata and attachments from an
existing mcap file. To add an attachment to a file,

    mcap add attachment demo.mcap -f Makefile

To get an attachment out of the file,

    mcap get attachment demo.mcap -n Makefile -o attachment.txt

To add metadata to a file,

    mcap add metadata demo.mcap -k foo=bar -k bar=baz -n "my metadata"

To get metadata out of the file,

    mcap get metadata demo.mcap -n "my metadata"
    {
      "bar": "baz",
      "foo": "bar"
    }

Also fixes a bug in the list attachments subcommand. Previously this was
listing chunks.