Add experimental optimize subcommand
This PR is for POC/discussion only.

Implements an "optimize" subcommand that accepts an mcap file and a list
of topics, aggregates the chosen topics from the input mcap file into a
second, in-memory mcap file, and attaches the second file to the end of
the first.

The idea of this is to enable efficient preloading of small topics for
display, regardless of the original size of the mcap. To exhibit the
functionality, select a large mcap file. For example, I have
cal_loop.mcap which is converted from cal_loop.bag. The file is 1.5 GB.

To simulate the current best performance for mcap preloading, run this
command with a selection of small topics:

    $ time mcap cat cal_loop.mcap --topics /diagnostics > /dev/null

    real    0m2.382s
    user    0m2.769s
    sys     0m0.408s

To "optimize" the file, run this command:

    $ time mcap optimize cal_loop.mcap -t /diagnostics > cal_loop.optimized.mcap

Now to simulate the optimized preloading performance, run this:

    $ time mcap get attachment cal_loop.optimized.mcap --name preload_topics | mcap cat > /dev/null

    real    0m0.010s
    user    0m0.017s
    sys     0m0.003s

Performance is 238x improved. Change the cat redirect to `--json` to
play the messages to stdout. Multiple topics may be supplied with
multiple -t flags.