Go: add support for stream handling of attachments
Adds support for streaming reads and writes of attachment data. With
this change, ParseAttachment is dropped as a public function and
replaced with a private function, parseAttachmentReader.

A new type AttachmentReader is exposed, which exposes the Data portion
of an attachment as a reader that can be streamed. The Attachment type
is modified to include a new ContentLength field, to allow the writer to
write the attachment length before stream processing the body.

A new option, AttachmentCallback, is exposed in the Lexer. Rather than
emitting attachments, the Lexer now parses an attachment reader and
calls the supplied callback (if supplied), and consumes any unconsumed
bytes from the attachment record (whether or not a callback is
supplied).

This is a breaking change in the Go SDK, and introduces some awkwardness
due to the presence of the new ContentLength field in the Attachment
record. In the specification, that value is folded into the Data field,
however this is not a tenable approach for stream handling. The actual
binary format is not affected by this - only the way in which the binary
format is described.

This change is required for readers and writers to be able to handle
attachments bigger than available RAM. Without it, attaching a
sufficiently large attachment will crash the reader/writer with an OOM.