Go: address goroutine leak when lexing zstd-compressed mcap (#749)
* Go: address goroutine leak when lexing zstd-compressed mcap

Adds Close() methods to the Lexer and Reader types. These are required
to avoid a goroutine leak when working with zstd-compressed files.

* Go: upgrade zstd library

This upgrades the zstd dependency, and actually fixes the goroutine
leak. I observe in the leak scenario that the zstd library leaks a
goroutine listening to this channel being delivered to, after closing
the reader:

https://github.com/klauspost/compress/blob/v1.14.1/zstd/blockdec.go#L604

I am still unsure what causes this leak to manifest under some run
conditions but not others. The problem goes away with the library
upgrade. I am continuing to investigate the bug and when it was fixed.