Update mcap dependencies in CLI tool
This should fix an installation error that occurs when installing via

    go install github.com/foxglove/mcap/go/cli/mcap@latest

When running the install locally the build succeeds; I think usage of go
workspaces is overriding the dependency specified in the go.mod file.
Need to understand the mechanics of that better to propose a long-term
fix to keep these in sync. For now this just updates the deps which I
expect will resolve the immediate problem.