Avoid divide by zero in info on subsecond inputs
Prior to this commit, if the duration of an mcap file were less than a
second we would divide by zero when computing the frequency on a topic.
This was because a float was getting cast to uint64 and getting
truncated before being used as the denominator in the frequency
computation.