Fix bug in ros1msg parser
Prior to this commit, if a ROS 1 message definition contained both
qualified and unqualified complex types in a particular order, it would
incorrectly infer the parent package of a succeeding unqualified type,
resulting in an erroneous failure to find a submessage definition. For a
concrete example, a message definition like this, parsed in the context
of package "original_package":

    Foo foo
    ===
    MSG: Foo
    other_package/Bar bar
    FooBar foobar
    ===
    MSG: other_package/Bar
    string data
    ===
    MSG: original_package/FooBar
    string data

would result in the parser attempting to look up other_package/FooBar
for FooBar's type definition, instead of looking for it in the parent
package of the overall type (i.e original_package).