Speed up ROS to JSON transcoder
This speeds up the ROS to JSON transcoder by removing dependency on the
encoding/json package from both it and the cat command.

Prior to this commit, JSON transcoding spent inordinate amounts of CPU
time compacting the JSON the transcoder was emitting, however that JSON
was almost compact already. Removing the encoder from the cat command
revealed that the encoding of strings out of the transcoder was not
compact - they got encoded as a quoted string followed by a newline.
Removing encoding/json from the transcoder and replacing it with
strconv.QuoteToAscii fixed this.