Docs: rearrange record ordering more logically
In the description of the summary section, swaps the location of
"metadata index" and "statistics". Files can be written in either order
technically, but this is what our libraries do. Metadata indexes are
also more logically grouped with attachment indexes.

In the listing of records, arranges the "data portion" records above the
summary section records".