Support fast addition of attachments/metadata
Changes the implementation of "mcap add attachment" and "mcap add
metadata" to append to the data section in place, then overwrite an
updated summary section with all offsets and CRCs updated.

This makes adding attachments and metadata very fast, and constant time
with respect to size of the file. Prior to this commit, adding an
attachment would rewrite the entire file, meaning the performance was
linear in the size of the file.