cli: fix segfault bug when using filter args (#1011)
### Public-Facing Changes

Fixes bug where the MCAP cli would segfault when using `filter -y` or
`filter -n`.
### Description

Previously the `compileMatchers` function would build an array of
matchers by appending to it, but the array was already initialized with
a non-zero length, resulting in the first N entries of that array being
uninitialized matchers. This caused a segfault when it came time to use
those matchers.
<!-- Link relevant Github issues. Use `Fixes #1234` to auto-close the
issue after merging. -->