Go: do not implicitly close custom decompressor in lexer
This removes some implicit behavior when using the lexer with a custom
decompressor. Prior to this commit, if your custom decompressor exposed
a Close() method it would get closed on close of the lexer. After this
commit, the caller is responsible for that. The benefit of this is they
will be better able to reuse their decompressor if desired.