Go: Fixes bug in metadata handling for indexed iteration
Prior to this commit, when using the "metadata callback" feature of the
indexed message iterator, the reader would mistakenly interpret a byte
as a "Token" rather than an "OpCode", causing the parsing of record
types to get confused.

This wasn't surfaced in the CLI metadata merging support because that
uses the unindexed iterator, which is not affected.

This commit fixes the bug and adds a test for both cases.