CLI: Append attachments to merged mcap files (#1052)
This updates `mcap merge` to include attachments from input files.

Attachments are appended in the order they are encountered, and may be
out of order with respect to log time. This procedure works differently
than metadata because of the way we process attachment opcodes.

- If an input file contains a summary statistics record and attachment
indexes, we scan any attachments from the those indexes
- If a summary statistics record contains an attachmentCount of 0, no
work is done
- Otherwise, for unindexed inputs, this scans the input file a second
time and processes attachments using the lexer's attachment callback.
This avoids changing any public interface NewReader.

We considered modifying the NewReader interface to accept an attachment
callback option, which would simplify the merging code (and avoid a
second scan on unindexed files), but deemed it better to leave the
interface alone.