spec: specify that DataEnd will not be extended (#1061)
### Public-Facing Changes

The spec has been updated to say that DataEnd records will not have extra records added in the future. This will not affect most readers or writers of MCAP files, but enables workflows for appending data to existing indexed files.

### Description

This change allows readers to assume that `Footer.summary_offset - 13`
is the beginning of the `DataEnd` record. Without it, this line of code
is wrong:
https://github.com/foxglove/mcap/blob/4d62967abe95e755b91f16cce390f6560d18e07e/go/cli/mcap/utils/mcap_amendment.go#L45-L46

See also: #1060