Go writer: support schemas/channels outside chunks
Adds a new option, ChannelsOutsideChunks, to the go writer.

Today in chunked MCAP files, schemas and channels are inside the chunks.
I would like to be able to concatenate two nonoverlapping MCAP files
without either decompressing chunks or having duplicate schemas/channels
in the output. This option should make it possible to do that.