mcap-protobuf-support: restore topological sort (#1162)
### Changelog

- Fixed: `mcap-protobuf-support` previously failed to decode Protobuf
messages where the message schema file descriptor set was not written in
topological dependency order.

### Description

#1148 changed the logic around loading file descriptor sets to decode
protobuf messages. The new logic is not able to handle protobuf schemas
where the file descriptors were not in topological dependency order.
Tests did not catch this because the `mcap_protobuf` python writer
writes file descriptors in topological order. However, the reader needs
to be able to handle different orders.

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->