spec: Clarify legal contents of chunk records
Prior to this commit, the wording of the spec implied that only channel,
schema, and message records are allowed in chunks. This could lead
tooling implementers to validate that assertion.

This is inconsistent with our support for extensible record types. Users
should be able to put custom records in chunks as well, and readers that
parse chunks should ignore unrecognized records rather than erroring on
them.