doctor: warn on duplicate matching schemas and channels (#1172)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->
- Added: `mcap doctor` now warns if a schema or channel record is
duplicated in the data section. Some writers do this and it wastes
space, though reading still works.
- Added: `mcap doctor` now prints an error if an MCAP contains messages
in indexed chunks where their channel and schema information are not
duplicated in the summary section. This is required by the spec:
https://mcap.dev/spec#chunk-index-op0x08


### Docs

Spec updated to clarify what channels and schemas must be repeated in
the summary section.

### Description


<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->