go: reader: read files with chunk indexes but not message indexes correctly. (#1179)
### Changelog
- Fixed: if reading an MCAP file written with chunk index records but no
message index records, the go MCAP reader would mistakenly yield no
messages.
- Fixed: the indexed MCAP reader would occasionally read too many bytes
when reading a chunk.
### Docs

None.

### Description

#### Message indexes
Right now the go indexed reader does not rely on message index records
for anything except to check if a chunk contains messages for that
topic. This is incorrect - if a chunk index does not contain message
index offsets, it can mean either:
- there are no messages in this chunk
- message indexes are not available for this chunk

See the spec: https://mcap.dev/spec#chunk-index-op0x08

#### chunk buffer sizing
- When resizing the chunk buffer between chunk reads, the reader grows
the chunk buffer by a multiple of the new chunk size if it's not big
enough. However, it would not correctly set the `len` for that buffer,
causing the next read to read more past the end of the chunk record into
the chunk buffer. This does not cause the reader to yield incorrect
information because we use use the contents of the chunk to determine
how long the compressed data in it is. However, it causes more bytes to
be read than neccessary. Also, it's possible to construct an MCAP where
this would cause an "unexpected EOF" error.