Fix table output with wide columns (#1191)
Fix: table output from list commands supports wide columns such as large
amounts of metadata

Previously, listing the metadata for an mcap file with many (>64kb)
key/value pairs failed silently.

The Scanner used in the table formatter has a buffer limit of 64kb, and
the error result was being ignored. This removes the scanner entirely
from most table output. It was introduced to trim leading whitespace
from lines, but the formatter accepts other options to make this happen.

I've given the info command its own formatter, which I think is
appropriate because it's formatting data differently, even including
tabs in its row data. The scanner's error result is now checked, though
we shouldn't see this in practice for `mcap info`.

This also makes a couple of copy fixes to the inline help for the
metadata commands.

Fixes #1189.