[rust] Support writing without chunks Topic: nochunk
On very memory-constrained systems like microcontrollers, the overhead of storing all the chunk
offset data in memory until the end of the file can be significant. To avoid that, we instead
write all Channel, Schema, and Message data directly as records rather than using chunks,
and rely on the reader (or an intermediate ingestion process) to generate the index information.