rust: Support writing without chunks (#1201)
On very memory-constrained systems like microcontrollers, the overhead
of storing all the chunk offset data in memory until the end of the file
can be significant. To avoid that, we instead write all Channel, Schema,
and Message data directly as records rather than using chunks, and rely
on the reader (or an intermediate ingestion process) to generate the
index information.

### Changelog
The rust library now (optionally) supports writing unchunked MCAP files
for decreased memory footprint.

### Docs

None

### Description

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

In our internal testing, storing chunk data required >40KB of memory for
several-minute logs, which is a lot for an embedded system with only
300KB of RAM total. Notably, this grows without bound during the test,
since the offsets and data about all chunks needs to be stored until the
end of the log when it will be written to the index.

For our usecase, it's acceptable to disable chunking entirely (along
with compression and indexing), since we can add those things later
during the log upload process.

We're using this internally on an ESP32-S3 and it is working well.

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->

---------

Co-authored-by: Eric Miller <eric@autopallet.bot>