python: add record size limit (#1225)
### Changelog
- Added: the python MCAP library reader classes gained a new parameter,
`record_size_limit`, defaulting to 4GiB. This allows callers to limit
the size of records that their application will support. When a record
is encountered with a greater length, the reader will raise an
`InvalidRecordLength` exception. This limit can be removed by setting it
to `None`. This helps applications avoid the issue where a corrupt MCAP
can cause a `MemoryError`.

### Docs

See generated python docs in the mcap.dev preview.
### Description

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->
Fixes: #1220