rust: add async reading functionality (#1211)
### Changelog
- rust: switches the LZ4 compression dependency from `lz4_flex` to
`lz4-rs`. This moves us from using a pure-rust lz4 implementation to C
bindings. I believe this is worthwhile because `lz4_flex` does not
support LZ4 "high compression mode". The practical reason for doing so
in this PR is that `lz4_flex` does not expose interfaces that make it
easy to build an AsyncRead adapter for it, but `lz4-rs` does.
- rust: Adds structs to read MCAP data asynchronously in a linear
stream.

### Docs

- Check generated rust docs for review.


### Description
Adds an async `RecordReader`implementation, for reading MCAP data
asynchronously. This is an optional feature, named `tokio`. I chose this
feature flag name and this module name because this functionality is
tied heavily into the Tokio ecosystem. If at some point we rebuild this
to be async-executor-agnostic, we can add that functionality under a new
module and feature flag name.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->